requirements.txt
absl-py==0.13.0
alabaster==0.7.12
albumentations==1.0.3
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
attrs==21.2.0
av==8.0.3
Babel==2.9.1
backcall==0.2.0
black==21.4b2
bleach==4.0.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
cloudpickle==1.6.0
commonmark==0.9.1
cycler==0.10.0
Cython==0.29.24
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
detectron2==0.5+cpu
docutils==0.16
efficientnet-pytorch==0.6.3
entrypoints==0.3
future==0.18.2
fvcore==0.1.5.post20210812
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
hydra-core==1.1.0
idna==3.2
imageio==2.9.0
imagesize==1.2.0
importlib-resources==5.2.2
iopath==0.1.8
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mock==4.0.3
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.1
omegaconf==2.1.0
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
packaging==21.0
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
portalocker==2.3.0
pretrainedmodels==0.7.4
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pydot==1.4.2
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.10.0
recommonmark==0.6.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scipy==1.7.1
segmentation-models-pytorch==0.2.0
Send2Trash==1.8.0
six==1.16.0
snowballstemmer==2.1.0
Sphinx==3.0.0
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.11.1
testpath==0.5.0
tifffile==2021.8.8
timm==0.4.12
toml==0.10.2
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.1
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yacs==0.1.8
zipp==3.5.0
